// src/app/shared/models/user-transaction.model.ts
export class UserTransaction {
  tx_id: string;
  type: string;
  type_title: string;
  type_shorttitle: string;
  user: string;
  transaction_dt: string;
  transaction_ts: string;
  currency: string;
  currency_sign: string;
  amount: string;
  amount_formatted: string;
  credit: string;
  credit_formatted: string;
  status: string;
  method_name: string;
  method_title: string;
  subtype_name: string;
  subtype_title: string;
  psp_title: string;
  comments: string;
  source_name: string;
  source_title: string;
  account_id: string;
}